+++ Tortuga.Chain/Tortuga.Chain.Core/Materializers/ConstructibleMaterializer`4.cs
@@ -13,9 +13,9 @@ namespace Tortuga.Chain.Materializers;
 /// <typeparam name="TObject">The type of the object that will be returned.</typeparam>
 /// <seealso cref="Materializer{TCommand, TParameter, TResult}"/>
 /// <seealso cref="IConstructibleMaterializer{TResult}"/>
-public abstract class ConstructibleMaterializer<TCommand, TParameter, TResult, TObject> : Materializer<TCommand, TParameter, TResult>, IConstructibleMaterializer<TResult>
-	where TCommand : DbCommand
-	where TParameter : DbParameter
+public abstract class ConstructibleMaterializer<TCommand, TParameter, TResult, TObject> : ColumnSelectingMaterializer<TCommand, TParameter, TResult>, IConstructibleMaterializer<TResult>
+where TCommand : DbCommand
+where TParameter : DbParameter
 {
 	/// <summary>
 	/// Initializes a new instance of the <see cref="ConstructibleMaterializer{TCommand,
@@ -32,18 +32,6 @@ public abstract class ConstructibleMaterializer<TCommand, TParameter, TResult, T
 	/// </summary>
 	/// <value>The data reader constructor.</value>
 	protected ConstructorMetadata? Constructor { get; set; }
-
-	/// <summary>
-	/// Columns to ignore when generating the list of desired columns.
-	/// </summary>
-	/// <value>The excluded columns.</value>
-	protected IReadOnlyList<string>? ExcludedColumns { get; private set; }
-
-	/// <summary>
-	/// Only include the indicated columns when generating the list of desired columns.
-	/// </summary>
-	/// <value>The included columns.</value>
-	protected IReadOnlyList<string>? IncludedColumns { get; private set; }
 
 	/// <summary>
 	/// Gets or sets the TObject metadata.
@@ -101,7 +89,7 @@ public abstract class ConstructibleMaterializer<TCommand, TParameter, TResult, T
 	/// Excludes the properties from the list of what will be populated in the object.
 	/// </summary>
 	/// <param name="propertiesToOmit">The properties to omit.</param>
-	public ILink<TResult> ExceptProperties(params string[] propertiesToOmit)
+	public override ILink<TResult> ExceptProperties(params string[] propertiesToOmit)
 	{
 		if (propertiesToOmit == null || propertiesToOmit.Length == 0)
 			return this;
@@ -303,7 +291,7 @@ public abstract class ConstructibleMaterializer<TCommand, TParameter, TResult, T
 	/// </summary>
 	/// <param name="propertiesToPopulate">The properties of the object to populate.</param>
 	/// <returns>ILink&lt;TResult&gt;.</returns>
-	public ILink<TResult> WithProperties(params string[] propertiesToPopulate)
+	public override ILink<TResult> WithProperties(params string[] propertiesToPopulate)
 	{
 		if (propertiesToPopulate == null || propertiesToPopulate.Length == 0)
 			return this;