+++ Tortuga.Chain/Tortuga.Chain.Core/Materializers/MultipleRow/DynamicCollectionMaterializer`2.cs
@@ -9,7 +9,8 @@ namespace Tortuga.Chain.Materializers;
 /// </summary>
 /// <typeparam name="TCommand">The type of the t command type.</typeparam>
 /// <typeparam name="TParameter">The type of the t parameter type.</typeparam>
-internal sealed class DynamicCollectionMaterializer<TCommand, TParameter> : Materializer<TCommand, TParameter, List<dynamic>> where TCommand : DbCommand
+internal sealed class DynamicCollectionMaterializer<TCommand, TParameter> : ColumnSelectingMaterializer<TCommand, TParameter, List<dynamic>>
+	where TCommand : DbCommand
 	where TParameter : DbParameter
 {
 	/// <summary>
@@ -19,17 +20,6 @@ internal sealed class DynamicCollectionMaterializer<TCommand, TParameter> : Mate
 	public DynamicCollectionMaterializer(DbCommandBuilder<TCommand, TParameter> commandBuilder)
 		: base(commandBuilder)
 	{ }
-
-	/// <summary>
-	/// Returns the list of columns the materializer would like to have.
-	/// </summary>
-	/// <returns>
-	/// IReadOnlyList&lt;System.String&gt;.
-	/// </returns>
-	/// <remarks>
-	/// If AutoSelectDesiredColumns is returned, the command builder is allowed to choose which columns to return. If NoColumns is returned, the command builder should omit the SELECT/OUTPUT clause.
-	/// </remarks>
-	public override IReadOnlyList<string> DesiredColumns() => AllColumns;
 
 	/// <summary>
 	/// Execute the operation synchronously.