+++ Tortuga.Chain/Tortuga.Chain.MySql/MySql/CommandBuilders/MySqlDeleteSet.cs
@@ -10,16 +10,9 @@ namespace Tortuga.Chain.MySql.CommandBuilders;
 /// <summary>
 /// Class MySqlDeleteSet.
 /// </summary>
-internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand, MySqlParameter>
+internal sealed class MySqlDeleteSet : DeleteSetDbCommandBuilder<MySqlCommand, MySqlParameter>
 {
-	readonly object? m_ArgumentValue;
-	readonly int? m_ExpectedRowCount;
-	readonly FilterOptions m_FilterOptions;
-	readonly object? m_FilterValue;
-	readonly DeleteOptions m_Options;
-	readonly IEnumerable<MySqlParameter>? m_Parameters;
 	readonly TableOrViewMetadata<MySqlObjectName, MySqlDbType> m_Table;
-	readonly string? m_WhereClause;
 
 	/// <summary>
 	/// Initializes a new instance of the <see cref="MySqlDeleteSet" /> class.
@@ -30,16 +23,12 @@ internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand,
 	/// <param name="parameters">The parameters.</param>
 	/// <param name="expectedRowCount">The expected row count.</param>
 	/// <param name="options">The options.</param>
-	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, string whereClause, IEnumerable<MySqlParameter> parameters, int? expectedRowCount, DeleteOptions options) : base(dataSource)
+	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, string whereClause, IEnumerable<MySqlParameter> parameters, int? expectedRowCount, DeleteOptions options) : base(dataSource, whereClause, parameters, expectedRowCount, options)
 	{
 		if (options.HasFlag(DeleteOptions.UseKeyAttribute))
 			throw new NotSupportedException("Cannot use Key attributes with this operation.");
 
 		m_Table = dataSource.DatabaseMetadata.GetTableOrView(tableName);
-		m_WhereClause = whereClause;
-		m_Parameters = parameters;
-		m_Options = options;
-		m_ExpectedRowCount = expectedRowCount;
 	}
 
 	/// <summary>
@@ -49,11 +38,9 @@ internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand,
 	/// <param name="tableName">Name of the table.</param>
 	/// <param name="whereClause">The where clause.</param>
 	/// <param name="argumentValue">The argument value.</param>
-	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, string whereClause, object? argumentValue) : base(dataSource)
+	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, string whereClause, object? argumentValue) : base(dataSource, whereClause, argumentValue)
 	{
 		m_Table = dataSource.DatabaseMetadata.GetTableOrView(tableName);
-		m_WhereClause = whereClause;
-		m_ArgumentValue = argumentValue;
 	}
 
 	/// <summary>
@@ -63,11 +50,9 @@ internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand,
 	/// <param name="tableName">Name of the table.</param>
 	/// <param name="filterValue">The filter value.</param>
 	/// <param name="filterOptions">The options.</param>
-	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, object filterValue, FilterOptions filterOptions) : base(dataSource)
+	public MySqlDeleteSet(MySqlDataSourceBase dataSource, MySqlObjectName tableName, object filterValue, FilterOptions filterOptions) : base(dataSource, filterOptions, filterOptions)
 	{
 		m_Table = dataSource.DatabaseMetadata.GetTableOrView(tableName);
-		m_FilterValue = filterValue;
-		m_FilterOptions = filterOptions;
 	}
 
 	public override CommandExecutionToken<MySqlCommand, MySqlParameter> Prepare(Materializer<MySqlCommand, MySqlParameter> materializer)
@@ -84,23 +69,23 @@ internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand,
 		if (sqlBuilder.HasReadFields)
 		{
 			sqlBuilder.BuildSelectClause(sql, "SELECT ", null, " FROM " + m_Table.Name.ToQuotedString());
-			if (m_FilterValue != null)
-				sql.Append(" WHERE " + sqlBuilder.ApplyFilterValue(m_FilterValue, m_FilterOptions));
-			else if (!string.IsNullOrWhiteSpace(m_WhereClause))
-				sql.Append(" WHERE " + m_WhereClause);
+			if (FilterValue != null)
+				sql.Append(" WHERE " + sqlBuilder.ApplyFilterValue(FilterValue, FilterOptions));
+			else if (!string.IsNullOrWhiteSpace(WhereClause))
+				sql.Append(" WHERE " + WhereClause);
 			sql.AppendLine(";");
 		}
 
 		sql.Append("DELETE FROM " + m_Table.Name.ToQuotedString());
-		if (m_FilterValue != null)
+		if (FilterValue != null)
 		{
-			sql.Append(" WHERE " + sqlBuilder.ApplyFilterValue(m_FilterValue, m_FilterOptions));
+			sql.Append(" WHERE " + sqlBuilder.ApplyFilterValue(FilterValue, FilterOptions));
 			parameters = sqlBuilder.GetParameters();
 		}
-		else if (!string.IsNullOrWhiteSpace(m_WhereClause))
+		else if (!string.IsNullOrWhiteSpace(WhereClause))
 		{
-			sql.Append(" WHERE " + m_WhereClause);
-			parameters = SqlBuilder.GetParameters<MySqlParameter>(m_ArgumentValue);
+			sql.Append(" WHERE " + WhereClause);
+			parameters = SqlBuilder.GetParameters<MySqlParameter>(ArgumentValue);
 			parameters.AddRange(sqlBuilder.GetParameters());
 		}
 		else
@@ -109,30 +94,15 @@ internal sealed class MySqlDeleteSet : MultipleRowDbCommandBuilder<MySqlCommand,
 		}
 		sql.Append(";");
 
-		if (m_Parameters != null)
-			parameters.AddRange(m_Parameters);
+		if (Parameters != null)
+			parameters.AddRange(Parameters);
 
-		return new MySqlCommandExecutionToken(DataSource, "Delete from " + m_Table.Name, sql.ToString(), parameters).CheckDeleteRowCount(m_Options, m_ExpectedRowCount);
+		return new MySqlCommandExecutionToken(DataSource, "Delete from " + m_Table.Name, sql.ToString(), parameters).CheckDeleteRowCount(Options, ExpectedRowCount);
 	}
 
 	/// <summary>
-	/// Returns the column associated with the column name.
+	/// Called when ObjectDbCommandBuilder needs a reference to the associated table or view.
 	/// </summary>
-	/// <param name="columnName">Name of the column.</param>
-	/// <returns></returns>
-	/// <remarks>
-	/// If the column name was not found, this will return null
-	/// </remarks>
-	public override ColumnMetadata? TryGetColumn(string columnName) => m_Table.Columns.TryGetColumn(columnName);
-
-	/// <summary>
-	/// Returns a list of columns known to be non-nullable.
-	/// </summary>
-	/// <returns>
-	/// If the command builder doesn't know which columns are non-nullable, an empty list will be returned.
-	/// </returns>
-	/// <remarks>
-	/// This is used by materializers to skip IsNull checks.
-	/// </remarks>
-	public override IReadOnlyList<ColumnMetadata> TryGetNonNullableColumns() => m_Table.NonNullableColumns;
+	/// <returns>TableOrViewMetadata.</returns>
+	protected override TableOrViewMetadata OnGetTable() => m_Table;
 }