+++ Tortuga.Chain/Tortuga.Chain.MySql/MySql/CommandBuilders/MySqlInsertBatch`1.cs
@@ -80,10 +80,14 @@ internal class MySqlInsertBatch<TObject> : DbCommandBuilder<MySqlCommand, MySqlP
 	/// <remarks>
 	/// If the column name was not found, this will return null
 	/// </remarks>
-	public override ColumnMetadata? TryGetColumn(string columnName)
-	{
-		return m_Table.Columns.TryGetColumn(columnName);
-	}
+	public override ColumnMetadata? TryGetColumn(string columnName) => m_Table.Columns.TryGetColumn(columnName);
+
+	/// <summary>
+	/// Returns a list of columns.
+	/// </summary>
+	/// <returns>If the command builder doesn't know which columns are available, an empty list will be returned.</returns>
+	/// <remarks>This is used by materializers to skip exclude columns.</remarks>
+	public override IReadOnlyList<ColumnMetadata> TryGetColumns() => m_Table.Columns;
 
 	/// <summary>
 	/// Returns a list of columns known to be non-nullable.