+++ Tortuga.Chain/Tortuga.Chain.MySql/MySql/CommandBuilders/MySqlTableOrView.cs
@@ -232,28 +232,10 @@ public class MySqlTableOrView<TObject> : TableDbCommandBuilder<MySqlCommand, MyS
 	}
 
 	/// <summary>
-	/// Returns the column associated with the column name.
+	/// Called when ObjectDbCommandBuilder needs a reference to the associated table or view.
 	/// </summary>
-	/// <param name="columnName">Name of the column.</param>
-	/// <returns></returns>
-	/// <remarks>
-	/// If the column name was not found, this will return null
-	/// </remarks>
-	public override ColumnMetadata? TryGetColumn(string columnName)
-	{
-		return m_Table.Columns.TryGetColumn(columnName);
-	}
-
-	/// <summary>
-	/// Returns a list of columns known to be non-nullable.
-	/// </summary>
-	/// <returns>
-	/// If the command builder doesn't know which columns are non-nullable, an empty list will be returned.
-	/// </returns>
-	/// <remarks>
-	/// This is used by materializers to skip IsNull checks.
-	/// </remarks>
-	public override IReadOnlyList<ColumnMetadata> TryGetNonNullableColumns() => m_Table.NonNullableColumns;
+	/// <returns>TableOrViewMetadata.</returns>
+	protected override TableOrViewMetadata OnGetTable() => m_Table;
 
 	/// <summary>
 	/// Adds (or replaces) the filter on this command builder.