+++ Tortuga.Chain/Tortuga.Chain.PostgreSql/PostgreSql/CommandBuilders/PostgreSqlInsertBulk.cs
@@ -65,6 +65,13 @@ namespace Tortuga.Chain.PostgreSql.CommandBuilders
 		public override ColumnMetadata? TryGetColumn(string columnName) => m_Table.Columns.TryGetColumn(columnName);
 
 		/// <summary>
+		/// Returns a list of columns.
+		/// </summary>
+		/// <returns>If the command builder doesn't know which columns are available, an empty list will be returned.</returns>
+		/// <remarks>This is used by materializers to skip exclude columns.</remarks>
+		public override IReadOnlyList<ColumnMetadata> TryGetColumns() => m_Table.Columns;
+
+		/// <summary>
 		/// Returns a list of columns known to be non-nullable.
 		/// </summary>
 		/// <returns>