+++ Tortuga.Chain/Tortuga.Chain.PostgreSql/PostgreSql/CommandBuilders/PostgreSqlSqlCall.cs
@@ -1,20 +1,15 @@
 ﻿using Npgsql;
-using System.Collections.Immutable;
 using Tortuga.Chain.CommandBuilders;
 using Tortuga.Chain.Core;
 using Tortuga.Chain.Materializers;
-using Tortuga.Chain.Metadata;
 
 namespace Tortuga.Chain.PostgreSql.CommandBuilders
 {
 	/// <summary>
 	/// Class PostgreSqlSqlCall
 	/// </summary>
-	public class PostgreSqlSqlCall : MultipleTableDbCommandBuilder<NpgsqlCommand, NpgsqlParameter>
+	public class PostgreSqlSqlCall : DbSqlCall<NpgsqlCommand, NpgsqlParameter>
 	{
-		readonly object? m_ArgumentValue;
-		readonly string m_SqlStatement;
-
 		/// <summary>
 		/// Initializes a new instance of the <see cref="PostgreSqlSqlCall"/> class.
 		/// </summary>
@@ -22,13 +17,10 @@ namespace Tortuga.Chain.PostgreSql.CommandBuilders
 		/// <param name="sqlStatement">The SQL statement.</param>
 		/// <param name="argumentValue">The argument value.</param>
 		/// <exception cref="ArgumentException">sqlStatement is null or empty.;sqlStatement</exception>
-		public PostgreSqlSqlCall(PostgreSqlDataSourceBase dataSource, string sqlStatement, object? argumentValue) : base(dataSource)
+		public PostgreSqlSqlCall(PostgreSqlDataSourceBase dataSource, string sqlStatement, object? argumentValue) : base(dataSource, sqlStatement, argumentValue)
 		{
 			if (string.IsNullOrEmpty(sqlStatement))
 				throw new ArgumentException($"{nameof(sqlStatement)} is null or empty.", nameof(sqlStatement));
-
-			m_SqlStatement = sqlStatement;
-			m_ArgumentValue = argumentValue;
 		}
 
 		/// <summary>
@@ -40,28 +32,7 @@ namespace Tortuga.Chain.PostgreSql.CommandBuilders
 		/// </returns>
 		public override CommandExecutionToken<NpgsqlCommand, NpgsqlParameter> Prepare(Materializer<NpgsqlCommand, NpgsqlParameter> materializer)
 		{
-			return new PostgreSqlCommandExecutionToken(DataSource, "Raw SQL Call", m_SqlStatement, SqlBuilder.GetParameters<NpgsqlParameter>(m_ArgumentValue));
+			return new PostgreSqlCommandExecutionToken(DataSource, "Raw SQL Call", SqlStatement, SqlBuilder.GetParameters<NpgsqlParameter>(ArgumentValue));
 		}
-
-		/// <summary>
-		/// Returns the column associated with the column name.
-		/// </summary>
-		/// <param name="columnName">Name of the column.</param>
-		/// <returns></returns>
-		/// <remarks>
-		/// If the column name was not found, this will return null
-		/// </remarks>
-		public override ColumnMetadata? TryGetColumn(string columnName) => null;
-
-		/// <summary>
-		/// Returns a list of columns known to be non-nullable.
-		/// </summary>
-		/// <returns>
-		/// If the command builder doesn't know which columns are non-nullable, an empty list will be returned.
-		/// </returns>
-		/// <remarks>
-		/// This is used by materializers to skip IsNull checks.
-		/// </remarks>
-		public override IReadOnlyList<ColumnMetadata> TryGetNonNullableColumns() => ImmutableList<ColumnMetadata>.Empty;
 	}
 }